+++ app/src/main/java/com/example/cs_2340_assignment_1/view/MainActivity.java
@@ -1,4 +1,4 @@
-package com.example.cs_2340_assignment_1;
+package com.example.cs_2340_assignment_1.view;
 
 import android.os.Bundle;
 import android.view.Menu;
@@ -10,6 +10,7 @@ import androidx.navigation.Navigation;
 import androidx.navigation.ui.AppBarConfiguration;
 import androidx.navigation.ui.NavigationUI;
 
+import com.example.cs_2340_assignment_1.R;
 import com.example.cs_2340_assignment_1.databinding.ActivityMainBinding;
 
 public class MainActivity extends AppCompatActivity {
