+++ app/src/main/res/layout/home_fragment.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Home">
+    tools:context=".view.Home">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/first_fragment"
