+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/Home"
-        android:name="com.example.cs_2340_assignment_1.Home"
+        android:name="com.example.cs_2340_assignment_1.view.Home"
         android:label="Home"
         tools:layout="@layout/home_fragment">
 
@@ -27,7 +27,7 @@
     </fragment>
     <fragment
         android:id="@+id/Courses"
-        android:name="com.example.cs_2340_assignment_1.CourseHome"
+        android:name="com.example.cs_2340_assignment_1.view.CourseHome"
         android:label="Courses"
         tools:layout="@layout/course_fragment">
 
@@ -40,7 +40,7 @@
     </fragment>
     <fragment
         android:id="@+id/AddCourse"
-        android:name="com.example.cs_2340_assignment_1.CourseAdd"
+        android:name="com.example.cs_2340_assignment_1.view.CourseAdd"
         android:label="Add Course"
         tools:layout="@layout/course_add_fragment" >
         <action
@@ -52,7 +52,7 @@
     </fragment>
     <fragment
         android:id="@+id/Assignments"
-        android:name="com.example.cs_2340_assignment_1.AssignmentHome"
+        android:name="com.example.cs_2340_assignment_1.view.AssignmentHome"
         android:label="Assignments"
         tools:layout="@layout/assignment_fragment" >
         <action
@@ -64,7 +64,7 @@
     </fragment>
     <fragment
         android:id="@+id/AddAssignment"
-        android:name="com.example.cs_2340_assignment_1.AssignmentAdd"
+        android:name="com.example.cs_2340_assignment_1.view.AssignmentAdd"
         android:label="Add Assignment"
         tools:layout="@layout/assignment_add_fragment" >
         <action
@@ -73,7 +73,7 @@
     </fragment>
     <fragment
         android:id="@+id/Exams"
-        android:name="com.example.cs_2340_assignment_1.ExamHome"
+        android:name="com.example.cs_2340_assignment_1.view.ExamHome"
         android:label="Exam"
         tools:layout="@layout/exam_fragment" >
         <action
@@ -85,7 +85,7 @@
     </fragment>
     <fragment
         android:id="@+id/AddExam"
-        android:name="com.example.cs_2340_assignment_1.ExamAdd"
+        android:name="com.example.cs_2340_assignment_1.view.ExamAdd"
         android:label="Add Exam"
         tools:layout="@layout/exam_add_fragment">
         <action
@@ -94,7 +94,7 @@
     </fragment>
     <fragment
         android:id="@+id/TodoList"
-        android:name="com.example.cs_2340_assignment_1.TodoHome"
+        android:name="com.example.cs_2340_assignment_1.view.TodoHome"
         android:label="Todo List"
         tools:layout="@layout/todo_fragment" >
         <action
@@ -106,7 +106,7 @@
     </fragment>
     <fragment
         android:id="@+id/AddTodoItems"
-        android:name="com.example.cs_2340_assignment_1.TodoAdd"
+        android:name="com.example.cs_2340_assignment_1.view.TodoAdd"
         android:label="Add Item"
         tools:layout="@layout/todo_add_fragment" >
         <action
